+++ packages/design-tokens/.storybook/main.js
@@ -17,5 +17,5 @@ module.exports = {
 			},
 		},
 	],
-	framework: '@storybook/html',
+	framework: '@storybook/react',
 };